requirements.txt
alabaster==0.7.16
Babel==2.14.0
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
docutils==0.21.2
Flask==3.0.3
Flask-SocketIO==5.3.6
h11==0.14.0
idna==3.7
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.3
lgpio==0.2.2.0
MarkupSafe==2.1.5
netifaces==0.10.6
packaging==24.0
Pygments==2.17.2
python-engineio==4.9.0
python-socketio==5.11.2
requests==2.31.0
rpi-lgpio==0.5
simple-websocket==1.0.0
snowballstemmer==2.2.0
socketio==0.2.1
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
urllib3==2.2.1
Werkzeug==3.0.2
wsproto==1.2.0

smbus2~=0.4.3
numpy~=1.26.4
opencv-python~=4.9.0.80